Do not fail on provided != MPI_THREAD_MULTIPLE
We had the same problem in the Table class, which upon initialization
would fail, as the automatic MPI initialization code assumed that
MPI_THREAD_MULTIPLE was the level of thread support required by the
user. In reality one simply doesn't know, but signaling an error is not
the way to go, as, for example, even simple tools like showtableinfo or
taql fail to read (in a non-mpirun, single-threaded manner) tables that
have been written using ADIOS2.

This code anyway should be put together at some point with that on the
Table.cc module.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>