Implementing column-wise scalar read/write I/O
This is yet a new set of I/O patterns to be offered by Adios2StMan.
Until now our implementation has been capable of handling scalar values,
but only on a per-cell basis. This means that if an application is
trying to read/write values from more than one row in a single pass,
this single call is translated into #rows calls into Adios2StMan. This
of course has big implications: reads and writes are slower than they
should be, as there is a bigger overhead due to the accumulation of
small operations; moreover, when writing data we hit again the
relatively high metadata-to-data ratio of ADIOS2, resulting in big
on-disk data sizes.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>