Exclude tests from build at configuration time
Until now, when indicated via -DBUILD_TESTING=OFF, tests have been
excluded from builds by excluding them from the "all" target. This
however has the disadvantage that their CMakeLists.txt files are still
processed at configuration time. While in theory this might not be a
problem, in reality it means adding potential complexity to a build that
will not need it.

As an example, in casa/Arrays/test a dependency (optional) is created on
Boost's unit testing framework. This dependency would need to be
satisfied at configuration time, even when the tests are not meant to be
built, adding a burden to users who just want a compiling system.

This commit changes the inclusion of all the test/ subdirectories from
being a build-time decision to being a configuration-time one. The
condition is still the same (BUILD_TESTING must be ON), so there should
be no change in behavior except from the intended simplification of
non-test builds.

Note that with this change we can now safely turn Boost's unit testing
framework for casa/Arrays/test into a hard dependency.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>