Make Boost's unit test framework REQUIRED
Now that tests are added conditionally at configuration time rather than
build time we can more safely make the dependency on Boost's unit
testing framework required rather than optional. Having it as optional
meant that unless the dependency wasn't satisfied all the array-related
tests would not get executed (and this was indeed the case in the Linux
Travis builds). However, given that the Array classes have been very
recently refactored it makes sense to force these tests to be built and
run alongside the rest of the casacore test suite.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>