Install Boost's unit testing framework
This is needed by the array tests, which previously were compiled
optionally only if this dependency was met, but now have been changed to
fail if it isn't.

The array unit tests not only depend on the unit testing framework
itself but also on Boost's system and filesystem libraries, so we need
to install those as well.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>